requirements.txt
# Core & Utilities
python-dotenv>=1.0.1
requests>=2.32.3
pathlib

# Data Processing
pandas>=2.2.3
numpy>=2.2.0
polars>=1.21.0
pyarrow>=19.0.0
scipy>=1.15.0

# Database
duckdb>=1.1.3

# Machine Learning
scikit-learn>=1.6.1
catboost>=1.2.7

# Visualization
plotly>=6.0.0

# App Framework
streamlit>=1.41.0